Restructured all files, tested code as well.
