+++ adstxt/adstxt/spiders/parser/parsers/helper.py
